% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oneway.R
\name{oneway}
\alias{oneway}
\title{One-way analysis of variance}
\usage{
oneway(formula, data)
}
\arguments{
\item{formula}{takes a formula for a linear model to find determine the relationship}

\item{data}{dataset necessary for analysis}
}
\value{
a list of analysis output
}
\description{
Input the formula to test the differences between means within groups of the data with
ANOVA technique.
}
\details{
takes the necessary formula for analyzing categorical and numeric data columns.
}
\examples{
\dontrun{
# analysis of variance in multiple categorical groups
anova <- oneway(mtcars$mpg ~ mtcars$cyl, mtcars)

#' # another example without $ format
anova <- oneway(mpg ~ cyl, mtcars)
 }
}
