% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.oneway}
\alias{plot.oneway}
\title{Plot one-way analysis of variance}
\usage{
\method{plot}{oneway}(x, col = "skyblue", ...)
}
\arguments{
\item{col}{sets color}

\item{...}{passes other attributives to graph properties}

\item{analysis}{takes a oneway(function) with desired data and attributes.}
}
\value{
list
}
\description{
Graphs a box plot for results of the one-way anova
}
\details{
creates a box plot which aids in the immediate comprehension of ANOVA
results in graph form.
}
\examples{
\dontrun{
# analysis of variance in multiple categorical groups
anova <- oneway(mtcars$mpg ~ mtcars$cyl, mtcars)
plot(anova)

#' # another example without $ format
anova <- oneway(mpg ~ cyl, mtcars)
plot(anova)
 }
}
