% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.oneway}
\alias{print.oneway}
\title{Print one-way analysis of variance}
\usage{
\method{print}{oneway}(x, ...)
}
\arguments{
\item{formula}{takes a oneway(function) with desired data and attributes.}
}
\value{
a list of data for plot which is rendered
}
\description{
Prints results of the one-way anova with additional formatting.
}
\details{
creates a list which aids in the immediate comprehension of ANOVA results in text form.
}
\examples{
\dontrun{
# analysis of variance in multiple categorical groups
anova <- oneway(mtcars$mpg ~ mtcars$cyl, mtcars)
print(anova)

#' # another example without $ format
anova <- oneway(mpg ~ cyl, mtcars)
print(anova)
 }
}
