% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.oneway}
\alias{summary.oneway}
\title{Summarize one-way analysis of variance}
\usage{
\method{summary}{oneway}(x, ...)
}
\arguments{
\item{analysis}{takes a oneway(function) with desired data and attributes.}
}
\value{
a list of data for plot which is rendered
}
\description{
Summarizes results of the one-way anova with additional formatting.
}
\details{
creates a list which aids in the immediate comprehension of ANOVA results in text form.
}
\examples{
\dontrun{
# analysis of variance in multiple categorical groups
anova <- oneway(mtcars$mpg ~ mtcars$cyl, mtcars)
summarize(anova)

#' # another example without $ format
anova <- oneway(mpg ~ cyl, mtcars)
summarize(anova)
 }
}
