% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{tar}
\alias{tar}
\alias{tar-package}
\title{R's Internal 'tar' Implementation as a Package}
\description{
There have been various issues with R's builtin 'tar' implementation
recently. This packages contains a version that fixes several of these
issues, so we can have a stable implementation that works with any R
version.
}
